+++ src/Hledger/Flow/Types.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE TypeSynonymInstances #-}
+{-# LANGUAGE FlexibleInstances #-}
+
 module Hledger.Flow.Types
 where
 
@@ -5,6 +8,8 @@ import Turtle
 import Prelude hiding (FilePath, putStrLn)
 
 data LogMessage = StdOut Text | StdErr Text | Terminate deriving (Show)
+type FullOutput = (ExitCode, Text, Text)
+type FullTimedOutput = (FullOutput, NominalDiffTime)
 
 class HasVerbosity a where
   verbose :: a -> Bool
@@ -17,3 +22,6 @@ class HasExitCode a where
 
 instance HasExitCode ExitCode where
   exitCode c = c
+
+instance HasExitCode FullOutput where
+  exitCode (c, _, _) = c